+++ edison-core/src/main/java/de/otto/edison/authentication/LdapAuthenticationFilter.java
@@ -26,7 +26,7 @@ import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 
 /**
  * Filter that checks for LDAP authentication once per request. Will not filter routes starting with
- * {@link LdapProperties#whitelistedPaths}. Uses {@link LdapProperties} to create an SSL based connection to the
+ * {@link LdapProperties#allowlistedPaths}. Uses {@link LdapProperties} to create an SSL based connection to the
  * configured LDAP server. Rejects requests with {@code HTTP 401} if authorization fails.
  */
 public class LdapAuthenticationFilter extends OncePerRequestFilter {
@@ -48,7 +48,7 @@ public class LdapAuthenticationFilter extends OncePerRequestFilter {
     @Override
     protected boolean shouldNotFilter(final HttpServletRequest request) throws ServletException {
         final String servletPath = request.getServletPath();
-        return ldapProperties.getWhitelistedPaths()
+        return ldapProperties.getAllowlistedPaths()
                 .stream()
                 .anyMatch(servletPath::startsWith);
     }