+++ CHANGELOG.md
@@ -305,7 +305,7 @@ Re-Enable Prefetch-Feature for `S3TogglzRepository` in newer Spring versions aga
 
 **New Features:**
 * **[edison-core]** Enhanced LDAP configuration to require a specific role to access secured paths. To achieve this you 
-                    can set the property `edison.ldap.required-role`. As usual, all whitelisted paths will be ignored.
+                    can set the property `edison.ldap.required-role`. As usual, all allowlisted paths will be ignored.
 * **[edison-core]** Enhanced LDAP configuration to use more than one prefix path with the property `edison.ldap.prefixes`.  
                     The old property `prefix` is now deprecated, but still usable and will be appended to `prefixes`.
                     
@@ -700,7 +700,7 @@ endpoints:
 
 **Bugfixes:**
 
-* **[edison-core]** Added /internal/js to LDAP whitelist, so the JS can be loaded w/o authentication in case of
+* **[edison-core]** Added /internal/js to LDAP allowlist, so the JS can be loaded w/o authentication in case of
 whilelisted /internal/jobs or /internal/jobdefinitions
 * **[edison-core]** Renamed Spring Bean `authenticationFilter` (see `LdapConfiguration`) to `ldapAuthenticationFilter`
 to prevent naming collisions with `authenticationFilter` bean registered by